fix: to the dashboard right after auth
