+++ src/app/dashboard/article/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from '@/components/layout';
+import Layout from '../../components/layout';
 const Appointment: React.FC = () => {
   return (
     <Layout>
