+++ src/app/dashboard/edit-profile/page.tsx
@@ -1,14 +1,13 @@
-import React from 'react';
-import Layout from '@/components/layout';
+import React from "react";
+import Layout from "../../components/layout";
 const Appointment: React.FC = () => {
   return (
     <Layout>
-   <div>
-      <h1 className="text-3xl font-bold">EDIT PROFILE</h1>
-      <p className="mt-4">Here you can manage your appointments.</p>
-    </div>
+      <div>
+        <h1 className="text-3xl font-bold">EDIT PROFILE</h1>
+        <p className="mt-4">Here you can manage your appointments.</p>
+      </div>
     </Layout>
- 
   );
 };
