+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import Page from "./pages";
 import About from "./pages/about";
 import Contact from "./pages/contact";
-import Footer from "../components/footer";
+import Footer from "./components/footer";
 
 
 export default function Home() {
